Move Inmotion and Ninebot into new packages
