+++ app/src/main/java/com/cooper/wheellog/MainActivity.kt
@@ -51,7 +51,11 @@ import com.cooper.wheellog.utils.PermissionsUtil.checkNotificationsPermissions
 import com.cooper.wheellog.utils.PermissionsUtil.isMaxBleReq
 import com.cooper.wheellog.utils.SomeUtil.getSerializable
 import com.cooper.wheellog.utils.SomeUtil.playBeep
+import com.cooper.wheellog.utils.inmotion.InMotionAdapter
+import com.cooper.wheellog.utils.inmotion.InmotionAdapterV2
 import com.cooper.wheellog.utils.kingsong.KingsongAdapter
+import com.cooper.wheellog.utils.ninebot.NinebotAdapter
+import com.cooper.wheellog.utils.ninebot.NinebotZAdapter
 import com.cooper.wheellog.views.PiPView
 import com.google.android.material.snackbar.Snackbar
 import com.welie.blessed.ConnectionState
