+++ app/src/main/java/com/cooper/wheellog/WheelData.java
@@ -10,6 +10,10 @@ import android.media.AudioManager;
 import com.cooper.wheellog.utils.*;
 import com.cooper.wheellog.utils.Constants.WHEEL_TYPE;
 import com.cooper.wheellog.utils.gotway.GotwayAdapter;
+import com.cooper.wheellog.utils.inmotion.InMotionAdapter;
+import com.cooper.wheellog.utils.inmotion.InmotionAdapterV2;
+import com.cooper.wheellog.utils.ninebot.NinebotAdapter;
+import com.cooper.wheellog.utils.ninebot.NinebotZAdapter;
 import com.cooper.wheellog.utils.veteran.VeteranAdapter;
 import com.cooper.wheellog.utils.kingsong.KingsongAdapter;
 import org.json.JSONArray;
