+++ app/src/main/java/com/cooper/wheellog/utils/ninebot/NinebotZAdapter.java
@@ -1,8 +1,12 @@
-package com.cooper.wheellog.utils;
+package com.cooper.wheellog.utils.ninebot;
 
 import com.cooper.wheellog.WheelData;
 import com.cooper.wheellog.WheelLog;
 import com.cooper.wheellog.R;
+import com.cooper.wheellog.utils.BaseAdapter;
+import com.cooper.wheellog.utils.MathsUtil;
+import com.cooper.wheellog.utils.SmartBms;
+import com.cooper.wheellog.utils.StringUtil;
 
 import java.io.ByteArrayOutputStream;
 import java.io.IOException;
